Optimize instance creation in __getattr__ [PREVIEW] (#210)
* Optimize instance creation in __getattr__ [PREVIEW]

Accessing a M2O field from a X2M field could result in as many
instantiations as there are records in the X2M field. It could be a
performance bottleneck due to the many creations of Model kwargs and
potential useless calls of transaction.set_context().

Thus the optimization is twofold:

 - Create an additional code path when the M2O field modifies the
   context
 - Cache the kwargs creation process to prevent extra calls resulting in
   the same dictionary

* Use read to fetch reverse field value of O2M field in get [PREVIEW]

Using a browse there in order to get the id is a performance bottleneck
as it will result in the instanciation of potentially thousands of
models.

---------

Co-authored-by: Jean Cavallo <jean.cavallo@hotmail.fr>