Return a promise from widgets display method
This allows for the scheduling of actions after the display of a view
has been finalized.

The display methods of the x2M widgets must take care of the fact that
the screen display being called afterwards should be return instead of
the maybe already resolve promise from the screen initial switch view.